Rename cssNodeContextLookup to nodeContextLookup
