+++ src/utils/index.js
@@ -2,7 +2,7 @@ export { default as blurComments } from "./blurComments"
 export { default as blurFunctionArguments } from "./blurFunctionArguments"
 export { default as configurationError } from "./configurationError"
 export { default as functionArgumentsSearch } from "./functionArgumentsSearch"
-export { default as cssNodeContextLookup } from "./cssNodeContextLookup"
+export { default as nodeContextLookup } from "./nodeContextLookup"
 export { default as blockString } from "./blockString"
 export { default as hasBlock } from "./hasBlock"
 export { default as hasEmptyBlock } from "./hasEmptyBlock"
